#458 fix: recover from wal will miss data
